backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
pymupdf
chromadb
asyncpg
sqlalchemy[asyncio]
pydantic
python-multipart
psycopg2-binary
fastapi[all] 